+++ tests/test_core.py
@@ -3,46 +3,13 @@ Performs general tests.
 """
 
 import numpy as np
-import pytest
-from pandas import DataFrame, Index, MultiIndex, Series
+from pandas import DataFrame, Index, MultiIndex
 from pandas.testing import assert_frame_equal, assert_index_equal
 
 from pandas_indexing.core import assignlevel
 
 
-@pytest.fixture
-def midx() -> MultiIndex:
-    return MultiIndex.from_tuples(
-        [("foo", 1), ("foo", 2), ("bar", 3)], names=["str", "num"]
-    )
-
-
-@pytest.fixture
-def sidx() -> Index:
-    return Index(["foo", "bar", "baz"], name="str")
-
-
-@pytest.fixture
-def mseries(midx) -> Series:
-    return Series([1, 2, 3], midx)
-
-
-@pytest.fixture
-def sseries(sidx) -> Series:
-    return Series([1, 2, 3], sidx)
-
-
-@pytest.fixture
-def sdf(sidx) -> DataFrame:
-    return DataFrame(dict(one=1, two=[1, 2, 3]), sidx)
-
-
-@pytest.fixture
-def mdf(midx) -> DataFrame:
-    return DataFrame(dict(one=1, two=[1, 2, 3]), midx)
-
-
-def test_assignlevel_index(sidx, midx):
+def test_assignlevel_index(sidx: Index, midx: MultiIndex):
     """
     Checks scalar setting of new and old levels.
     """
@@ -89,7 +56,7 @@ def test_assignlevel_index(sidx, midx):
     )
 
 
-def test_assignlevel_dataframe(mdf):
+def test_assignlevel_dataframe(mdf: DataFrame):
     """
     Checks setting dataframes on both axes.
     """